chore: Update font configuration with FiraCode
This commit updates the font configuration in the `common/configuration.nix` file to include the FiraCode font. Additionally, the `common/home/kitty.nix` file is modified to set the `font_family` to "FiraCode". This change improves the consistency and aesthetics of the font across the system.

Note: The commit message has been modified to adhere to the established commit message conventions.