feat: Update monitor and scaling settings in configuration.nix
This commit updates the monitor and scaling settings in the `common/configuration.nix` file. The "monitor" setting is changed to "highres" and the scaling is set to 2. Additionally, the `xwayland` configuration is modified to force zero scaling. The "GDK_SCALE" environment variable is set to 2 and the "XCURSOR_SIZE" is set to 32. These changes enhance the display configuration and improve the user experience.